+++ lang/syn/src/parser/accounts/mod.rs
@@ -14,12 +14,11 @@ pub fn parse(strct: &syn::ItemStruct) -> ParseResult<AccountsStruct> {
     let instruction_api: Option<Punctuated<Expr, Comma>> = strct
         .attrs
         .iter()
-        .filter(|a| {
+        .find(|a| {
             a.path
                 .get_ident()
                 .map_or(false, |ident| ident == "instruction")
         })
-        .next()
         .map(|ix_attr| ix_attr.parse_args_with(Punctuated::<Expr, Comma>::parse_terminated))
         .transpose()?;
     let fields = match &strct.fields {
@@ -68,11 +67,16 @@ pub fn parse_account_field(f: &syn::Field, has_instruction_api: bool) -> ParseRe
 }
 
 fn is_field_primitive(f: &syn::Field) -> ParseResult<bool> {
-    let r = match ident_string(f)?.as_str() {
-        "ProgramState" | "ProgramAccount" | "CpiAccount" | "Sysvar" | "AccountInfo"
-        | "CpiState" | "Loader" => true,
-        _ => false,
-    };
+    let r = matches!(
+        ident_string(f)?.as_str(),
+        "ProgramState"
+            | "ProgramAccount"
+            | "CpiAccount"
+            | "Sysvar"
+            | "AccountInfo"
+            | "CpiState"
+            | "Loader"
+    );
     Ok(r)
 }
 
@@ -113,12 +117,12 @@ fn ident_string(f: &syn::Field) -> ParseResult<String> {
 }
 
 fn parse_program_state(path: &syn::Path) -> ParseResult<ProgramStateTy> {
-    let account_ident = parse_account(&path)?;
+    let account_ident = parse_account(path)?;
     Ok(ProgramStateTy { account_ident })
 }
 
 fn parse_cpi_state(path: &syn::Path) -> ParseResult<CpiStateTy> {
-    let account_ident = parse_account(&path)?;
+    let account_ident = parse_account(path)?;
     Ok(CpiStateTy { account_ident })
 }
 
@@ -161,20 +165,16 @@ fn parse_account(path: &syn::Path) -> ParseResult<syn::Ident> {
                     let path_segment = &ty_path.path.segments[0];
                     Ok(path_segment.ident.clone())
                 }
-                _ => {
-                    return Err(ParseError::new(
-                        args.args[1].span(),
-                        "first bracket argument must be a lifetime",
-                    ))
-                }
+                _ => Err(ParseError::new(
+                    args.args[1].span(),
+                    "first bracket argument must be a lifetime",
+                )),
             }
         }
-        _ => {
-            return Err(ParseError::new(
-                segments.arguments.span(),
-                "expected angle brackets with a lifetime and type",
-            ))
-        }
+        _ => Err(ParseError::new(
+            segments.arguments.span(),
+            "expected angle brackets with a lifetime and type",
+        )),
     }
 }